requirements.txt
xgboost
absl-py==0.9.0
alembic==1.4.2
astor==0.8.1
attrdict==2.0.1
attrs==19.3.0
bravado==10.6.0
bravado-core==5.17.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cliff==3.1.0
cmaes==0.5.0
cmd2==0.8.9
colorlog==4.1.0
cycler==0.10.0
future==0.18.2
gast==0.2.2
gitdb==4.0.4
GitPython==3.1.1
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
joblib==0.14.1
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
neptune-client==0.4.111
neptune-contrib==0.19.0
neptune-lib==0.1.2
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
optuna==1.3.0
pandas==1.0.3
pbr==5.4.5
Pillow==7.1.2
prettytable==0.7.2
protobuf==3.11.3
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
smmap==3.0.2
SQLAlchemy==1.3.16
stevedore==1.32.0
strict-rfc3339==0.7
swagger-spec-validator==2.5.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.45.0
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.1.9
webcolors==1.11.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
